combine classes, conversations, and ielts into one routes file
