+++ src/pages/ClassesPage.jsx
@@ -4,8 +4,8 @@ import { useLocation, useParams, Link } from 'wouter';
 import Class from '../components/Class/Class';
 import Level from '@/components/Class/Level';
 import SkeletonClass from '@/components/Skeletons/SkeletonClass';
-import { getClasses } from '@/wrapper/class-wrapper';
-import { getLevels } from '@/wrapper/level-wrapper';
+import { getClasses } from '@/wrappers/class-wrapper';
+import { getLevels } from '@/wrappers/level-wrapper';
 import { useTranslation } from "react-i18next";
 import SkeletonLevel from '../components/Skeletons/SkeletonLevel';
 import Skeleton from 'react-loading-skeleton';
