+++ src/pages/dashboards/InstructorEditClass.jsx
@@ -2,11 +2,11 @@ import { useContext, useEffect, useState } from "react";
 import { UserContext } from '@/contexts/UserContext.jsx';
 import { useLocation, useParams } from 'wouter';
 import { useAuth } from '@clerk/clerk-react';
-import { getClassById } from "@/wrapper/class-wrapper";
+import { getClassById } from "@/wrappers/class-wrapper";
 import FormInput from '@/components/Form/FormInput'
 import Button from '@/components/Button/Button';
-import { updateClass } from '@/wrapper/class-wrapper.js';
-import { getUser } from '@/wrapper/user-wrapper.js';
+import { updateClass } from '@/wrappers/class-wrapper.js';
+import { getUser } from '@/wrappers/user-wrapper.js';
 import BackButton from "@/components/Button/BackButton";
 import UserItem from "@/components/UserItem";
 import SkeletonUser from "@/components/Skeletons/SkeletonUser";
