+++ src/pages/dashboards/admin/AdminStudents.jsx
@@ -3,9 +3,9 @@ import { UserContext } from '@/contexts/UserContext.jsx';
 import { useLocation, Link } from 'wouter';
 import { useAuth } from '@clerk/clerk-react';
 import { IoPersonOutline } from "react-icons/io5";
-import { getUsers, getStudentsForExport } from '@/wrapper/user-wrapper.js';
-import { getClasses, getClassById } from '@/wrapper/class-wrapper';
-import { getStudentsClasses } from '@/wrapper/user-wrapper';
+import { getUsers, getStudentsForExport } from '@/wrappers/user-wrapper.js';
+import { getClasses, getClassById } from '@/wrappers/class-wrapper';
+import { getStudentsClasses } from '@/wrappers/user-wrapper';
 import Unauthorized from "@/pages/Unauthorized";
 import Dropdown from '@/components/Dropdown/Dropdown';
 import Button from '@/components/Button/Button';
