+++ src/pages/dashboards/admin/editPages/AddConversation.jsx
@@ -7,7 +7,7 @@ import Button from '@/components/Button/Button';
 import DayDropdown from '@/components/Dropdown/DayDropdown';
 import BackButton from "@/components/Button/BackButton";
 import Alert from '@/components/Alert';
-import { createConversation } from '@/wrapper/conversation-wrapper.js';
+import { createConversation } from '@/wrappers/conversation-wrapper.js';
 import { IoAdd, IoTrashBinOutline } from "react-icons/io5";
 import Unauthorized from "@/pages/Unauthorized";
