+++ src/pages/dashboards/admin/editPages/EditClass.jsx
@@ -10,8 +10,8 @@ import DayDropdown from '@/components/Dropdown/DayDropdown';
 import UserItem from "@/components/UserItem";
 import Alert from '@/components/Alert';
 import { IoAdd, IoTrashBinOutline, IoPersonOutline } from "react-icons/io5";
-import { updateClass, deleteClass, getClasses } from '@/wrapper/class-wrapper';
-import { getUser } from '@/wrapper/user-wrapper';
+import { updateClass, deleteClass, getClasses } from '@/wrappers/class-wrapper';
+import { getUser } from '@/wrappers/user-wrapper';
 import { convertTime } from "@/utils/time-utils";
 import Unauthorized from "@/pages/Unauthorized";
 import SkeletonUser from "@/components/Skeletons/SkeletonUser";
