+++ src/pages/dashboards/admin/editPages/EditIelts.jsx
@@ -2,14 +2,14 @@ import { useContext, useEffect, useState } from "react";
 import { UserContext } from '@/contexts/UserContext.jsx';
 import { useLocation, useParams } from 'wouter';
 import { useAuth } from '@clerk/clerk-react';
-import { getIeltsById } from "@/wrapper/ielts-wrapper";
+import { getIeltsById } from "@/wrappers/ielts-wrapper";
 import FormInput from '@/components/Form/FormInput'
 import Button from '@/components/Button/Button';
 import DeleteButton from "@/components/Button/DeleteButton";
 import DayDropdown from '@/components/Dropdown/DayDropdown';
 import BackButton from "@/components/Button/BackButton";
 import Alert from '@/components/Alert';
-import { updateIelts, deleteIelts } from '@/wrapper/ielts-wrapper.js';
+import { updateIelts, deleteIelts } from '@/wrappers/ielts-wrapper.js';
 import { IoAdd, IoTrashBinOutline } from "react-icons/io5";
 import Unauthorized from "@/pages/Unauthorized";
