+++ src/pages/dashboards/admin/editPages/EditLevel.jsx
@@ -3,8 +3,8 @@ import { UserContext } from '@/contexts/UserContext.jsx';
 import { useLocation, useParams } from 'wouter';
 import { useAuth } from '@clerk/clerk-react';
 import { useTranslation } from "react-i18next";
-import { getLevels, updateLevel, deleteLevel } from '@/wrapper/level-wrapper.js';
-import { getClasses, deleteClass } from '@/wrapper/class-wrapper.js';
+import { getLevels, updateLevel, deleteLevel } from '@/wrappers/level-wrapper.js';
+import { getClasses, deleteClass } from '@/wrappers/class-wrapper.js';
 import Button from '@/components/Button/Button';
 import DeleteButton from "@/components/Button/DeleteButton";
 import BackButton from "@/components/Button/BackButton";
