+++ src/pages/dashboards/admin/editPages/EditUser.jsx
@@ -2,8 +2,8 @@ import { useContext, useEffect, useState } from "react";
 import { UserContext } from '@/contexts/UserContext.jsx';
 import { useLocation, useParams } from 'wouter';
 import { useAuth } from '@clerk/clerk-react';
-import { updateUser, getUser, deleteUser } from '@/wrapper/user-wrapper.js';
-import { getClassById, getClasses, enrollInClass, unenrollInClass } from '@/wrapper/class-wrapper';
+import { updateUser, getUser, deleteUser } from '@/wrappers/user-wrapper.js';
+import { getClassById, getClasses, enrollInClass, unenrollInClass } from '@/wrappers/class-wrapper';
 import FormInput from '@/components/Form/FormInput';
 import PhoneInput from '@/components/Form/PhoneInput/PhoneInput';
 import Button from '@/components/Button/Button';
