+++ src/wrappers/conversation-wrapper.js
@@ -3,7 +3,7 @@ import { toTitleCase } from '@/utils/formatters';
 
 const getConversations = async () => {
   try {
-    const response = await axios.get("/api/conversations/");
+    const response = await axios.get("/api/classes/conversations/");
     return response.data;
   } catch (error) {
     console.error('Error fetching conversations:', error)
@@ -13,7 +13,7 @@ const getConversations = async () => {
 
 const getConversationById = async (conversationId) => {
   try {
-    const response = await axios.get(`/api/conversations/${conversationId}`);
+    const response = await axios.get(`/api/classes/conversations/${conversationId}`);
     return response.data;
   } catch (error) {
     console.error("Error fetching conversation from id:", error);
@@ -24,7 +24,7 @@ const getConversationById = async (conversationId) => {
 const createConversation = async (conversationData) => {
   conversationData.instructor = toTitleCase(conversationData.instructor);
   try {
-    const response = await axios.post('/api/conversations', conversationData);
+    const response = await axios.post('/api/classes/conversations', conversationData);
     return response.data;
   } catch (error) {
     console.error('Error updating conversation:', error);
@@ -37,7 +37,7 @@ const updateConversation = async (conversationId, conversationData) => {
     conversationData.instructor = toTitleCase(conversationData.instructor);
   }
   try {
-    const response = await axios.put(`/api/conversations/${conversationId}`, conversationData);
+    const response = await axios.put(`/api/classes/conversations/${conversationId}`, conversationData);
     return response.data;
   } catch (error) {
     console.error('Error updating conversation:', error);
@@ -47,7 +47,7 @@ const updateConversation = async (conversationId, conversationData) => {
 
 const deleteConversation = async (conversationId) => {
   try {
-    const response = await axios.delete(`/api/conversations/${conversationId}`);
+    const response = await axios.delete(`/api/classes/conversations/${conversationId}`);
     return response.data;
   } catch (error) {
     console.error('Error deleting conversation:', error);
